Move '/ip' endpoint for Api to InfoController. Add '/version' endpoint to EventsProcessor service
