+++ src/Api/Controllers/InfoController.cs
@@ -1,4 +1,6 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using Microsoft.AspNetCore.Mvc;
 using Bit.Core.Utilities;
 
@@ -12,11 +14,25 @@ namespace Bit.Api.Controllers
         {
             return DateTime.UtcNow;
         }
-        
+
         [HttpGet("~/version")]
         public JsonResult GetVersion()
         {
             return Json(CoreHelpers.GetVersion());
+        }
+        
+        [HttpGet("~/ip")]
+        public JsonResult Ip()
+        {
+            var headerSet = new HashSet<string> { "x-forwarded-for", "cf-connecting-ip", "client-ip" };
+            var headers = HttpContext.Request?.Headers
+                .Where(h => headerSet.Contains(h.Key.ToLower()))
+                .ToDictionary(h => h.Key);
+            return new JsonResult(new
+            {
+                Ip = HttpContext.Connection?.RemoteIpAddress?.ToString(),
+                Headers = headers,
+            });
         }
     }
 }
