+++ src/EventsProcessor/Startup.cs
@@ -49,7 +49,12 @@ namespace Bit.EventsProcessor
             app.UseEndpoints(endpoints =>
             {
                 endpoints.MapGet("/alive", 
-                    async context => await context.Response.WriteAsync(System.DateTime.UtcNow.ToString()));
+                    async context => await context.Response.WriteAsJsonAsync(System.DateTime.UtcNow));
+                endpoints.MapGet("/now",
+                    async context => await context.Response.WriteAsJsonAsync(System.DateTime.UtcNow));
+                endpoints.MapGet("/version",
+                    async context => await context.Response.WriteAsJsonAsync(CoreHelpers.GetVersion()));
+
             });
         }
     }
